requirements.txt
openai
tiktoken
langchain
Pillow
gradio
clip-interrogator
